Remember the search filters used by the user.
When a user searches their data, Lumidex should persist the filters used so the user doesn't have to select the filters again on next startup. Often times the same filters are used and it is annoying having to reselect the same ones every time.

Each filter's parameters are persisted to the database on application shutdown. The parameters are restored on startup.